Use more static methods
I saw an opportunity to use more static methods, so I changed a few
things around so that collaborators now mostly send messages from a
static context. I think it makes for a cleaner API in most cases, and
the APIs of these classes are so small that if it turns out it would be
better for things to be more literally object-oriented, I think it will
still be achievable.