Make ACF field group location args dynamic
Allow the ACF field groups defined in `acf/php/**` to define the ACF
location conditions dynamically. Effectively, the generated PHP code
contains a hard-coded path, but we need the path to be set dynamically.

Also, update the `README.md` to explain why this path needs to be set
dynamically, and how the behavior is implemented. For more information
about this, take a look at the documentation I added – the section is
called "Working with ACF".