Only import necessary parts of Bootstrap
Replace the blanket import of all Bootstrap SCSS and JS files with
specific imports. Avoid importing any part of Bootstrap that is not
being used.

This is specifically useful because it avoids a CSS conflict that was
observed when v1.0.0 of this plugin was installed in production. The
`.btn` Bootstrap class conflicted with a custom `.btn` CSS class defined
by the Trinity Rep WordPress theme. This plugin was previously importing
the `_button.scss` Bootstrap SCSS file, despite not needing it – by
eliminating the import of this unneeded file, the CSS conflict is no
longer an issue.

Generally speaking, I should have taken this step before releasing v1.0
of this plugin. Bad practice to ship any code that isn't necessary.